add a page for managing bookings
